Move more config out of the root, split things
SpringMVC is for UI, not just for Security
