+++ SpiNNaker-allocserv/src/main/java/uk/ac/manchester/spinnaker/alloc/web/SpallocServiceAPI.java
@@ -19,8 +19,8 @@ package uk.ac.manchester.spinnaker.alloc.web;
 import static io.swagger.v3.oas.annotations.enums.ParameterIn.PATH;
 import static javax.ws.rs.core.MediaType.APPLICATION_JSON;
 import static javax.ws.rs.core.MediaType.TEXT_PLAIN;
-import static uk.ac.manchester.spinnaker.alloc.SecurityConfig.IS_READER;
-import static uk.ac.manchester.spinnaker.alloc.SecurityConfig.IS_USER;
+import static uk.ac.manchester.spinnaker.alloc.security.SecurityConfig.IS_READER;
+import static uk.ac.manchester.spinnaker.alloc.security.SecurityConfig.IS_USER;
 import static uk.ac.manchester.spinnaker.alloc.web.DocConstants.T_JOB;
 import static uk.ac.manchester.spinnaker.alloc.web.DocConstants.T_MCH;
 import static uk.ac.manchester.spinnaker.alloc.web.DocConstants.T_TOP;
